lots of debug code

+++ trunk/src/vm/frame.h
@@ -29,7 +29,7 @@
 
 struct frame_struct
 {
-  uint32_t *vaddr;
+  uint8_t *vaddr;
   size_t length;
   uint32_t flag;
   block_sector_t sector_no;
@@ -52,7 +52,7 @@ struct pte_shared
 
 void sup_pt_init (void);
 
-bool sup_pt_add (uint32_t *pd, void *upage, uint32_t *vaddr, size_t length,
+bool sup_pt_add (uint32_t *pd, void *upage, uint8_t *vaddr, size_t length,
                  uint32_t flag, block_sector_t sector_no);
 bool sup_pt_shared_add (uint32_t *pd, void *upage, struct frame_struct *fs);
 
@@ -71,14 +71,14 @@ bool sup_pt_set_memory_map (uint32_t *pte, void *kpage);
 void sup_pt_fs_set_dirty (struct frame_struct *fs, bool dirty);
 bool sup_pt_fs_is_dirty  (struct frame_struct *fs);
 
-void sup_pt_fs_set_pte_list (struct frame_struct *fs, uint32_t *kpage,
+void sup_pt_fs_set_pte_list (struct frame_struct *fs, uint8_t *kpage,
                              bool is_swapping_in);
 
 bool sup_pt_fs_scan_and_reset_access (struct frame_struct *fs);
 
-uint32_t *sup_pt_evict_frame (void);
+uint8_t *sup_pt_evict_frame (void);
 
-bool mark_page (void *upage, uint32_t *addr, int length, uint32_t flag,
+bool mark_page (void *upage, uint8_t *addr, size_t length, uint32_t flag,
                 block_sector_t sector_no);
 
 #endif /* vm/frame.h */
